tdd_django_restfull_api/requirements.txt
# Django
django==1.11.0
djangorestframework==3.6.2


# Interaction with Postgres via Python
psycopg2==2.7.1